Fix to generate output which Prometheus will parse.
Fixes the time attached to metrics to be in the correct format for Prometheus.

Fixes the key names in dimensions to be able to deal with - in key names,
which is used / needed by some fullerite collectors.

Fix the issues with outputting a type statement for fullerite_collector_datapoints
multiple times by just throwing away that metric. Prometheus has it's own meta metrics

Fix logging as error to log as warn instead - this still shows up in the default
log level that we run with, but avoids triggering the logging hook and registering
collector errors.